fw-5518, move query modules into queries package
