+++ firstvoices/backend/search/queries/search_term_query.py
